requirements.txt
attrs==17.4.0
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
Django==2.0.1
django-chartjs-engine==0.0.4
django-filter==1.1.0
djangorestframework==3.7.7
idna==2.6
Markdown==2.6.11
pdfminer3k==1.3.1
pluggy==0.6.0
ply==3.10
psycopg2==2.7.3.2
py==1.5.2
pytest==3.4.0
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
